Rename bqrs-result to bqrs-raw-results-mapper
