+++ extensions/ql-vscode/src/common/bqrs-raw-results-mapper.ts
